Begin refactoring simple mesher for other changes (to test them before writing greedy mesher)

+++ terra-core/src/main/java/com/ritualsoftheold/terra/material/TerraTexture.java
@@ -76,7 +76,7 @@ public class TerraTexture {
         return texCoordY;
     }
     
-    public float getTexCoordZ() {
+    public int getTexCoordZ() {
         return texCoordZ;
     }
